\name{bcellViper-package}
\Rdversion{1.4}
\alias{bcellViper-package}
\alias{bcellViper}
\docType{package}
\title{
Human B-cell interactome and normal and tumor B-cell phenotypes expresion data
}
\description{
This package contains an interactome (regulon), expression data (dset), a list of normal sample IDs (normalSamples) and a short example of the ARACNE algorithm output (bcellaracne.adj file) required to run the examples of package viper.
}
\details{
\tabular{ll}{
Package: \tab bcellViper\cr
Type: \tab Package\cr
Version: \tab 0.99.2\cr
Date: \tab 2013-04-15\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Mariano Javier Alvarez

Maintainer: Mariano Javier Alvarez <malvarez@c2b2.columbia.edu>
}
\references{
Basso, K. et al. Reverse engineering of regulatory networks in human B cells. Nature genetics 37, 382-90 (2005).

Gautier, L., Cope, L., Bolstad, B. M., and Irizarry, R. A. 2004. affy---analysis of Affymetrix GeneChip data at the probe level. Bioinformatics 20, 3 (Feb. 2004), 307-315.

Alvarez, M. J., Sumazin, P., Rajbhandari, P. & Califano, A. Correlating measurements across samples improves accuracy of large-scale expression profile experiments. Genome biology 10, R143 (2009).
}

